+++ packages/froxel-gl/src/main.js
@@ -5,7 +5,7 @@ export { default as ShaderProgram } from './shader-program.js';
 export { default as Buffer } from './buffer.js';
 export { default as VertexBuffer } from './vertex-buffer.js';
 export { default as ElementBuffer } from './element-buffer.js';
-export { default as UniformBuffer } from './uniform-buffer.js';
+export { default as UniformBlockBuffer } from './uniform-block-buffer.js';
 
-export { default as VertexArrayObject } from './vertex-array-object.js';
-export { default as TextureObject } from './texture-object.js';
+export { default as VertexArray } from './vertex-array.js';
+export { default as Texture } from './texture.js';