+++ packages/froxel-gl/src/vertex-buffer.js
@@ -5,7 +5,7 @@ import Buffer from './buffer.js';
 export default class VertexBuffer extends Buffer
 {
 	/**
-	 * Creates a VertexBuffer linked to the ARRAY_BUFFER target.
+	 * Creates a buffer for the ARRAY_BUFFER target.
 	 * @param {WebGLCanvas} gl
 	 * @param {number} usage
 	 */